% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{BackwardLinear}
\alias{BackwardLinear}
\title{Adapted Backward Algorithm}
\usage{
BackwardLinear(data, time, tran, class, max2, max3)
}
\arguments{
\item{data}{two dimensional matrix for individual.  1st dimension is time 2nd is partially observed data}

\item{time}{what time to caluclate for.  Works recursivly, if time = t calculates for all values greater than t}

\item{tran}{transition probabilities}

\item{class}{classification probabilities}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}
}
\value{
list (each individual) of list (each time) of list (each partially observed data) of vector (forward probabilities)
}
\description{
Adapted backward that incorporates partially observed data.  Partial data is resolved in linear time.
}
