% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{CalcLikelihoodMover}
\alias{CalcLikelihoodMover}
\title{Calculates Likelihood of Movers}
\usage{
CalcLikelihoodMover(forw, pi)
}
\arguments{
\item{forw}{forward quantity for every individual}

\item{pi}{probability of being a stayer}
}
\value{
mover likelihood
}
\description{
Calculates likelihood of movers by summing over forward quantity
}
