% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Classification}
\alias{Classification}
\title{Classifies observed given truth}
\usage{
Classification(x_val, y_val, class)
}
\arguments{
\item{x_val}{observed value}

\item{y_val}{true value}

\item{class}{classification matrix}
}
\description{
Returns classification probability of x_val as y_val given classification probability
}
