% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{CombineandPattern}
\alias{CombineandPattern}
\title{Combines three tests and patternizes them}
\usage{
CombineandPattern(mat1_pers, mat2, mat3, max2, max3, reduce, time_length)
}
\arguments{
\item{mat1_pers}{3 dimensional data array of test 1}

\item{mat2}{matrix for test 2}

\item{mat3}{matrix for test 3}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{reduce}{for StateCalculator if true, if the first two tests are negative imputes negative for the third test in Combiner}

\item{time_length}{Number of observations per individual}
}
\value{
(list of patternized data and frequency vector)
}
\description{
Combines three tests and patternizes them.  Matrix 2 and 3 must be of the same size.  Matrix 1 must be an array with
    the first two dimensions matching those of matrix 2 and 3 and with a third dimension of atleast 2.
}
