% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Combiner}
\alias{Combiner}
\title{Combines three data matrices}
\usage{
Combiner(mat1, mat2, mat3, max2, max3, reduce)
}
\arguments{
\item{mat1}{first source data.  Must be array with a third dimension of at least 2.  3rd dimension can be all NAs.  Incorporated in CombineandPattern}

\item{mat2}{second data matrix}

\item{mat3}{third data matrix}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{reduce}{T/F that is used as reduce_ng in StateCalculator}
}
\value{
combined array
}
\description{
Combines three data matrices to one array
}
