% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{CompleteDataLogLikelihood}
\alias{CompleteDataLogLikelihood}
\title{Expected Value of Complete Data Log Likelihood Given First Lambda Value}
\usage{
CompleteDataLogLikelihood(
  lambda1,
  data_pattern,
  forward,
  backward,
  likelihoods,
  init,
  tran,
  class,
  pi_0
)
}
\arguments{
\item{lambda1}{First lambda value, since lambda has to sum to 1, can only optimize one and calculate the other easily}
}
\description{
Calculates the expected value of the complede data log likelihood given the first lambda value.  Used to optimize lambda given all other parameters
}
