% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Runs EM Algorithm}
\usage{
EM(
  data_array,
  freq_vec,
  epsilon,
  init,
  tran,
  class,
  pi_0,
  max2,
  max3,
  time_length
)
}
\arguments{
\item{data_array}{data array of three combined tests, patternized}

\item{freq_vec}{frequency vector of patterns}

\item{epsilon}{threshold for EM convergence, stops when likelihood percentage increase is below epsilon}

\item{init}{initial probabilities}

\item{tran}{transition probabilities}

\item{class}{classification probabilities}

\item{pi_0}{probability of being stayer}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{time_length}{Number of observations per individual}
}
\value{
list of initial parameters, estimated parameters and likelihood
    initial and estimated parameter entries have same format
    1) initial 2) transition 3) classification 4) stayer probability
}
\description{
Runs EM using adapted forward-backward that calculates partially observed data.
   Calculates probability of being a stayer, intiial vector, transition matrix, and classification matrix
   Uses the following functions: ForwardLinear, BackwardLinear, CalcLikelihoodMover, CalcLikelihoodStayer,
       CalcInitialLin, CalcTransitionLin, CalcClassificationLin, and CalcStayerLin
}
