% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EMRaff}
\alias{EMRaff}
\title{Runs EM Algorithm for Rafferty Method}
\usage{
EMRaff(
  data_pattern,
  freq_vec,
  epsilon,
  time_length,
  init,
  tran,
  class,
  pi_0,
  lambda,
  max2,
  max3
)
}
\arguments{
\item{freq_vec}{frequency vector of patterns}

\item{epsilon}{threshold for EM convergence, stops when likelihood percentage increase is below epsilon}

\item{time_length}{Number of observations per individual}

\item{init}{initial probabilities}

\item{tran}{transition probabilities}

\item{class}{classification probabilities}

\item{pi_0}{probability of being stayer}

\item{lambda}{vector of length two, first value is weight given to value at time t-1, second value is weight given to value at time t-2}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{data_array}{data array of three combined tests, patternized}
}
\value{
list of initial parameters, estimated parameters and likelihood
    initial and estimated parameter entries have same format
    1) initial 2) transition 3) classification 4) stayer probability
}
\description{
Runs EM using adapted forward-backward that calculates partially observed data for Rafferty method.
   Calculates probability of being a stayer, intiial vector, transition matrix, and classification matrix
   Uses the following functions: ForwardLinearRaff, BackwardLinearRaff, CalcLikelihoodMover, CalcLikelihoodStayer,
       CalcInitialLinRaff, CalcTransitionLin,Raff CalcClassificationLinRaff, and CalcStayerLin
}
