% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{ForwardLinearRaff}
\alias{ForwardLinearRaff}
\title{Adapted Forward Algorithm for Partially Observed Data and Rafferty Method}
\usage{
ForwardLinearRaff(data, time, init, tran, class, lambda, max2, max3)
}
\arguments{
\item{data}{two dimensional matrix for individual.  1st dimension is time 2nd is partially observed data}

\item{time}{what time to caluclate for.  Works recursivly, if time = t calculates for all values less than t}

\item{init}{initial probabilities}

\item{tran}{transition probabilities}

\item{class}{classification probabilities}

\item{lambda}{vector of length two, first value is weight given to value at time t-1, second value is weight given to value at time t-2}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}
}
\value{
list (each individual) of list (each time) of list (each partially observed data) of vector (forward probabilities)
}
\description{
Adapted forward that incorporates partially observed data.  Partial data is resolved in linear time. Also incorporates Rafferty Method as cited in paper.
Assumes transition rate is the same when considering time t-1 and t-2 however allows for different weights (lambda)
}
