% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GenerateData}
\alias{GenerateData}
\title{Generates Simulated Latent Data}
\usage{
GenerateData(n, t, p)
}
\arguments{
\item{n}{number of individuals to generate data for}

\item{t}{number of visits to generate data for}

\item{p}{probability of being a stayer in the healthy state for latent data}
}
\value{
list of true data and vector of indicators for whether each individual is a stayer
}
\description{
Generates simulated latent data from initial and transition probabilities.  Incorporated in GenerateSimulatedData()
}
