% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GenerateDataRaff}
\alias{GenerateDataRaff}
\title{GenerateData() for Rafferty Method}
\usage{
GenerateDataRaff(n, t, p, lambda)
}
\arguments{
\item{n}{number of individuals to generate data for}

\item{t}{number of visits to generate data for}

\item{p}{probability of being a stayer in the healthy state for latent data}

\item{lambda}{vector of length two, first value is weight given to value at time t-1, second value is weight given to value at time t-2}
}
\value{
list of true data and vector of indicators for whether each individual is a stayer
}
\description{
Generates simulated latent data from initial,transition and lambda probabilities.  Incorporated in GenerateSimulatedDataRaff()
}
