% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GenerateObs}
\alias{GenerateObs}
\title{Generates Simulated Observed Data}
\usage{
GenerateObs(data_true, max2, max3)
}
\arguments{
\item{data_true}{simulated latent data that error is added to}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}
}
\value{
list of observed data and true classification matrix
}
\description{
Generates simulated observed data from simulated latent data by adding measurment error.  Incorporated in GenerateSimulatedData()
}
