% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GenerateSimulatedData}
\alias{GenerateSimulatedData}
\title{Generates Simulated Data}
\usage{
GenerateSimulatedData(n, t, p, max2, max3)
}
\arguments{
\item{n}{number of individuals to generate}

\item{t}{number of timepoints to generate}

\item{p}{probability of being a stayer for simulated data}

\item{max2}{number of possibilities from test 2}

\item{max3}{number of possibilities from test}
}
\value{
list of three matrices, one for each test
}
\description{
Generates simulated HPV, cytology and colposcopy data by using the following functions:
   GenerateData, GenerateObs, GetClass, GetInitTrue, GetTranTrue, IntroduceMissing, IntroduceMissingAll, IntroduceMissingColpo, and Seperator
}
