% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GenerateSimulatedDataRaff}
\alias{GenerateSimulatedDataRaff}
\title{Generates Simulated Data for Rafferty Method}
\usage{
GenerateSimulatedDataRaff(n, t, p, lambda, max2 = 3, max3 = 2)
}
\arguments{
\item{n}{number of individuals to generate}

\item{t}{number of timepoints to generate}

\item{p}{probability of being a stayer for simulated data}

\item{lambda}{vector of length two, first value is weight given to value at time t-1, second value is weight given to value at time t-2}

\item{max2}{number of possibilities from test 2}

\item{max3}{number of possibilities from test}
}
\value{
list of three matrices, one for each test
}
\description{
Generates simulated HPV, cytology and colposcopy data by using the following functions:
   GenerateDataRaff, GenerateObs, GetClass, GetInitTrueRaff, GetTranTrueRaff, IntroduceMissing, IntroduceMissingAll, IntroduceMissingColpo, and Seperator
}
