% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GetInitTrue}
\alias{GetInitTrue}
\title{Calculates True Intial Probabilities}
\usage{
GetInitTrue(data_true, stayer_vec_true, max2, max3)
}
\arguments{
\item{data_true}{simulated latent data}

\item{stayer_vec_true}{vector of latent indicators for whether an individual is a stayer}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}
}
\value{
true initial vector
}
\description{
Calculates True Intial Probabilities from simulated latent data.  Incorporated in GenerateSimulatedData()
}
