% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{GetPatterns}
\alias{GetPatterns}
\title{Patternizes Data}
\usage{
GetPatterns(states)
}
\arguments{
\item{states}{large data array}
}
\value{
dataframe of patterns and frequencies.
}
\description{
Gets all patterns in data. Incorporated in CombineandPattern
}
