% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{IntroduceMissing}
\alias{IntroduceMissing}
\title{Introduces Misisng data}
\usage{
IntroduceMissing(ind_data_obs, p)
}
\arguments{
\item{ind_data_obs}{data to introduce independent randomness to}

\item{p}{probability that each observation is missing}
}
\value{
data matrix with missing values
}
\description{
Introduces missingness that is independent of the value of the other two tests.  Incorporated in GenerateSimulatedData()
}
