% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{IntroduceMissingAll}
\alias{IntroduceMissingAll}
\title{Introduced Missing Data}
\usage{
IntroduceMissingAll(mat1, mat2, mat3, p_time)
}
\arguments{
\item{mat1}{test 1}

\item{mat2}{test 2}

\item{mat3}{test 2}

\item{p_time}{vector of length t for missing probabilities for each time.}
}
\value{
list of three matrices with concurrent missing data
}
\description{
Introduces missing data to all three tests at the same time point to mimic mixed visits. Incorporated in GenerateSimulatedData()
}
