% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{IntroduceMissingColpo}
\alias{IntroduceMissingColpo}
\title{Introduces Misisng data}
\usage{
IntroduceMissingColpo(mat1, mat2, mat3)
}
\arguments{
\item{mat1}{first dependent test}

\item{mat2}{second dependent test}

\item{mat3}{test that misssingness is added to}
}
\value{
data matrix with missing values
}
\description{
Introduces missingness that is dependent on the value of the other two tests.  Incorporated in GenerateSimulatedData()
}
