% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{IntroducePersistence}
\alias{IntroducePersistence}
\title{Introduces Persistence}
\usage{
IntroducePersistence(hpv)
}
\arguments{
\item{hpv}{binary data matrix where one-obersvation back persistence is introduced}
}
\value{
time dependent trinary data matrix
}
\description{
Introduces one observation persistence to binary test.  Adds time dependence while still maintaining Markov identity.  Incorporated in GenerateSimulatedData.
}
