% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Normalize}
\alias{Normalize}
\title{Normalized Matrices}
\usage{
Normalize(data)
}
\arguments{
\item{data}{matrix that will be row normalied}
}
\value{
row normalized matrix
}
\description{
Normalizes matrices so rows sum to one
}
