% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{numOfStates}
\alias{numOfStates}
\title{Returns Number of states}
\usage{
numOfStates(max2, max3)
}
\arguments{
\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}
}
\value{
vector from 0 to k where k is the max value of the states
}
\description{
Returns Number of states
}
