% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Pattern2Data}
\alias{Pattern2Data}
\title{Turns patterns into data array}
\usage{
Pattern2Data(unique_patterns, max2, max3, time_length)
}
\arguments{
\item{unique_patterns}{patterns output from GetPatterns}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{time_length}{number of observations per person}
}
\value{
list of array of unique patterns and frequency vector
}
\description{
Turns patterns into data array.  Incorporated in CombineandPattern
}
