% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{persitenceHelper}
\alias{persitenceHelper}
\title{Helper Function for IntroducePersistence}
\usage{
persitenceHelper(val, one_back)
}
\arguments{
\item{val}{current value that is being made time dependent}

\item{one_back}{value from last time point}
}
\value{
time dependent value
}
\description{
Helper Function for IntroducePersistence
}
