% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{ProdClass}
\alias{ProdClass}
\title{Calculates Probability Latent Values are 0 given observed data}
\usage{
ProdClass(data, class)
}
\arguments{
\item{data}{dimensional matrix for individual.  1st dimension is time 2nd is partially observed data}

\item{class}{classification matrix}
}
\value{
probability of obsevred values if latent values are all 0
}
\description{
Calculates probability for individuals observed values assuming latent values are zero
}
