% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Seperator}
\alias{Seperator}
\title{Seperates Three Tests}
\usage{
Seperator(data_obs)
}
\arguments{
\item{data_obs}{simulated observed latent data}
}
\value{
list of three data matrices
}
\description{
Takes in data of simulated latent data of three tests and seperates them into three seperate matrices.  Incorporated in GenerateSimulatedData()
}
