% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{StateCalculator}
\alias{StateCalculator}
\title{Calculates State Given Individual Test Results}
\usage{
StateCalculator(val1, val2, val3, max2, max3, reduce_ng = F)
}
\arguments{
\item{val1}{value from test 1}

\item{val2}{value from test 2}

\item{val3}{value from test 3}

\item{max2}{number of values the second test can take on}

\item{max3}{number of values the third test can take on}

\item{reduce_ng}{if true, if the first two tests are negative imputes negative for the third test}
}
\value{
tuple value, may be vector if partially observed
}
\description{
Calculates state tuple result given individual tests, can output partially observed data. Incorportated in CombineandPattern().
}
